requirements.txt
fastapi[standard]
openai>=1.0.0
python-dotenv
